Remove obsolete CS RPM GPG key during IPU
Similarly to RHEL upgrades, we want to remove obsolete GPG keys, e.g.
keys using SHA1. On CS9 (and also CS8) the gpg-pubkey-8483c65d-5ccc5b19
key imported from /etc/pki/rpm-gpg/RPM-GPG-KEY-centosofficial is used.
This key uses SHA1 and makes CS9 -> CS10 IPUs fail due to SHA1 being
prohibited.

This patch adds the key to the list of keys to remove during IPU.

NOTE: The key used on CS10 is
/etc/pki/rpm-gpg/RPM-GPG-KEY-centosofficial-SHA256 which uses SHA256,
but after importing it it's also named gpg-pubkey-8483c65d-5ccc5b19.
This isn't a problem because the SHA1 one is removed before the SHA256
one is imported, but it's something to keep in mind if there is
CS10 -> CS11 sometime in he future.