networkmanagerconnectionscanner: Fix tests
Apparently the networkmanagerconnectionscanner tests never actually ran,
because in test containers the NetworkManager-libnm package was not
being installed and locally it seems it's not being picked up in
virtualenv.

When the package is installed, the test_nm_conn tests fails because
there is a bug in it.

There is an incorrect assertion about the number of properties in the
[connection] section of the test connection file (4 instead of 3) as
well an incorrect assumption about the keys within section being left in
the original order. However, it seems like either GLib's keyfile or
nmlib modify the order of the keys.

E.g. the test file has the keys in the connection section in this order:
type, id, wifi; but after the processing it's id, uuid, type.

This patch fixes the tests and also enables installation of
NetworkManager-libnm.
Also the NetworkManager-libnm is now installed in test containers:
- rhel9 (used in Makefile) - the tests run and pass
- ubi9 (used in CI) - the tests run and pass
- rhel8 (Makefile) - the tests are skipped because the package is not
  being picked up, not sure why
- ubi8 (CI) - the package is not available in the UBI repos, so not
  installed