Add --target-os CLI option
The option specifies the target OS (distribution) to upgrade to. This
sets the value of LEAPP_TARGET_OS, however LEAPP_DEVEL_TARGET_OS has
precedence. If none of the envars are defined, default to the source
distro i.e. only upgrade, no conversion, will be performed.

The available options are 'rhel', 'centos' and 'almalinux'.
Note that the "ID" value from /etc/os-release is used and therefore
'centos' really refers to Centos Stream.